% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_data.R
\name{gather_data}
\alias{gather_data}
\title{Gather Data}
\usage{
gather_data()
}
\value{
A data frame on stock-by-date information.
}
\description{
gather_data() brings together all the daily return data we need from the ws.data package.
}

